requirements.txt
python-telegram-bot>=13.5
pyTelegramBotAPI>=0.0.1
gspread>=3.7.0
oauth2client>=4.1.3
pytz